smart_farming website/requirements.txt
# requirements.txt
Flask==2.3.2
mysql-connector-python==8.3.0
Flask-CORS==4.0.0
openai==1.14.3
python-dotenv==1.0.1
matplotlib==3.7.1
